update tests/Fix Lint errors
