--- frontend/src/index.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import { BrowserRouter } from 'react-router-dom';
-import { createRoot } from 'react-dom/client';
-import App from './App';
-import { Provider } from 'react-redux';
-import store from './store/store';
-
-const appNode = createRoot(document.getElementById('ModernMan'));
-
-appNode.render(
-    <React.StrictMode>
-    <Provider store={store}>
-        <BrowserRouter>
-            <App />  
-        </BrowserRouter>
-    </Provider>
-    </React.StrictMode>
-);
