+++ frontend/webpack.prod.js
@@ -9,5 +9,4 @@ module.exports = merge(common, {
       chunks: 'all',
     },
   },
- },
-);
+});
